+++ tournament/src/tournament.rs
@@ -137,7 +137,7 @@ impl Tournament
                         "Time Elapsed: {:.2}h, Number of Games Played: {}, GPM: {:.2}",
                         starting_time.elapsed().as_secs() as f32 / (60 * 60) as f32,
                         num_of_games_played,
-                        num_of_games_played as f32 / starting_run_time.elapsed().as_secs() as f32 * 60 as f32
+                        num_of_games_played as f32 / starting_run_time.elapsed().as_secs() as f32 * 60_f32
                     );
 
                     presistance.write_game(&game_result)?;
